+++ main/gpu/src/test.cu
@@ -22,7 +22,7 @@ int main(int argc, char *argv[]) {
 	Ran ran(seed);
 	Point* points = new Point[n];
 	for (int i=0; i<n; ++i) {
-		double x, y;
+		float x, y;
 		ran.circle(x, y);
 
 		points[i].x[0] = x;
@@ -40,8 +40,8 @@ int main(int argc, char *argv[]) {
 //	Point* square = new Point[n*n];
 //	for (int i=0; i<n; ++i) {
 //		for (int j=0; j<n; ++j) {
-//			square[i*n + j].x[0] = (double)i/(double)n;
-//			square[i*n + j].x[1] = (double)j/(double)n;
+//			square[i*n + j].x[0] = (float)i/(float)n;
+//			square[i*n + j].x[1] = (float)j/(float)n;
 //		}
 //	}
 //
